chore(tests): Reorganize test files
