+++ src/rest-api/tests/features/emails.rs
@@ -9,12 +9,12 @@ use regex::Regex;
 use crate::TestWorld;
 
 #[then(expr = "{int} email(s) should have been sent")]
-fn then_n_emails_sent(world: &mut TestWorld, n: usize) {
+pub fn then_n_emails_sent(world: &mut TestWorld, n: usize) {
     assert_eq!(world.email_notifier_state().send_count, n);
 }
 
 #[then(expr = "the email body should match {string}")]
-fn then_email_matches(world: &mut TestWorld, pattern: Regex) {
+pub fn then_email_matches(world: &mut TestWorld, pattern: Regex) {
     let email = world.email_notifier_state().sent[0].message_plain.clone();
     assert!(pattern.is_match(&email));
 }
