+++ src/rest-api/tests/features/mod.rs
@@ -3,8 +3,10 @@
 // Copyright: 2024–2025, Rémi Bardon <remi@remibardon.name>
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
+pub mod app_config;
 pub mod auth;
 pub mod dns_setup;
+pub mod emails;
 pub mod init;
 pub mod invitations;
 pub mod members;
@@ -15,27 +17,27 @@ pub mod roles;
 pub mod server_config;
 #[cfg(feature = "test")]
 pub mod user_limit;
+pub mod util;
 pub mod workspace_details;
 
-pub(crate) mod prelude {
-    pub(crate) use std::str::FromStr as _;
+mod prelude {
+    pub(super) use std::str::FromStr as _;
 
-    pub(crate) use axum::http::{header::*, StatusCode};
-    pub(crate) use axum_test::{TestResponse, TestServer};
-    pub(crate) use base64::{prelude::BASE64_STANDARD, Engine as _};
-    pub(crate) use chrono::{TimeDelta, Utc};
-    pub(crate) use cucumber::{given, then, when};
-    pub(crate) use prose_pod_api::error::Error;
-    pub(crate) use secrecy::{ExposeSecret as _, SecretString};
-    pub(crate) use serde_json::json;
-    pub(crate) use service::{
+    pub(super) use axum::http::{header::*, StatusCode};
+    pub(super) use axum_test::{TestResponse, TestServer};
+    pub(super) use base64::{prelude::BASE64_STANDARD, Engine as _};
+    pub(super) use chrono::{TimeDelta, Utc};
+    pub(super) use cucumber::{given, then, when};
+    pub(super) use prose_pod_api::error::Error;
+    pub(super) use secrecy::{ExposeSecret as _, SecretString};
+    pub(super) use serde_json::json;
+    pub(super) use service::{
         errors::DbErr,
         models::xmpp::*,
         sea_orm::{prelude::*, IntoActiveModel as _, Set},
         MutationError,
     };
 
-    pub(crate) use crate::{
-        api_call_fn, cucumber_parameters as parameters, user_token, util::*, TestWorld,
-    };
+    pub(super) use crate::{api_call_fn, user_token, util::*};
+    pub(super) use crate::{cucumber_parameters as parameters, TestWorld};
 }
