+++ src/rest-api/tests/features/util/mod.rs
@@ -1,0 +1,8 @@
+// prose-pod-api
+//
+// Copyright: 2024–2025, Rémi Bardon <remi@remibardon.name>
+// License: Mozilla Public License v2.0 (MPL v2.0)
+
+pub mod http;
+pub mod lifecycle;
+pub mod sse;
