+++ src/rest-api/tests/prelude/mod.rs
@@ -3,10 +3,11 @@
 // Copyright: 2024–2025, Rémi Bardon <remi@remibardon.name>
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
+pub mod app_config;
 pub mod database;
 pub mod mocks;
-pub mod notifier;
-pub mod steps;
 pub mod test_api;
 pub mod test_world;
 pub mod util;
+
+pub use test_world::TestWorld;
