--- src/rest-api/tests/prelude/steps.rs
@@ -1,254 +1,0 @@
-// prose-pod-api
-//
-// Copyright: 2024–2025, Rémi Bardon <remi@remibardon.name>
-// License: Mozilla Public License v2.0 (MPL v2.0)
-
-use std::collections::HashSet;
-
-use cucumber::gherkin::Step;
-use lazy_static::lazy_static;
-use regex::Regex;
-use tracing::error;
-
-use crate::{features::prelude::*, test_api::test_server};
-
-pub mod app_config {
-    use std::sync::Arc;
-
-    use service::{
-        dependencies,
-        licensing::LicenseService,
-        secrets::{LiveSecretsStore, SecretsStore},
-        AppConfig,
-    };
-
-    use crate::{
-        features::prelude::*,
-        prelude::{
-            mocks::{MockLicenseService, MockSecretsStore},
-            test_world::CONFIG_PATH,
-        },
-    };
-
-    pub fn reload_config(world: &mut TestWorld) {
-        let figment =
-            AppConfig::figment_at_path(CONFIG_PATH.as_path()).merge(world.config_overrides.clone());
-        let config = AppConfig::from_figment(figment)
-            .expect(&format!("Invalid config file at {}", CONFIG_PATH.display()));
-
-        let mock_license_service = Arc::new(MockLicenseService::new(config.server_fqdn()));
-        let mock_secrets_store =
-            Arc::new(MockSecretsStore::new(LiveSecretsStore::default(), &config));
-        let uuid_gen = dependencies::Uuid::from_config(&config);
-
-        world.app_config = Some(Arc::new(config));
-        world.license_service = Some(LicenseService::new(mock_license_service.clone()));
-        world.mock_license_service = Some(mock_license_service);
-        world.secrets_store = Some(SecretsStore::new(mock_secrets_store.clone()));
-        world.mock_secrets_store = Some(mock_secrets_store);
-        world.uuid_gen = Some(uuid_gen);
-    }
-
-    #[given(expr = "config {string} is set to {}")]
-    fn given_app_config(world: &mut TestWorld, key: String, value: String) -> anyhow::Result<()> {
-        if world.api.is_some() {
-            tracing::warn!("Config set after the API has started, you need to restart.");
-            world.api = None;
-        }
-        world.set_config(&key, &value)?;
-        reload_config(world);
-
-        Ok(())
-    }
-
-    #[given(expr = "config {string} is unset")]
-    fn given_app_config_unset(world: &mut TestWorld, key: String) {
-        if world.api.is_some() {
-            tracing::warn!("Config set after the API has started, you need to restart.");
-            world.api = None;
-        }
-        world.unset_config(&key);
-        reload_config(world);
-    }
-}
-
-#[given(expr = "the Prose Pod API {state_verb} started")]
-async fn given_api_run_state(world: &mut TestWorld, state: parameters::StateVerb) {
-    if state.into_bool() {
-        assert!(world.api.is_none());
-        world.api = Some(test_server(world).await.unwrap());
-        world.reset_server_ctl_counts();
-    } else {
-        assert!(world.api.is_none());
-        world.api = None;
-    }
-}
-
-#[given(expr = "the Prose Pod API has restarted")]
-async fn given_api_restarted(world: &mut TestWorld) {
-    world.api = Some(test_server(world).await.unwrap());
-    world.reset_server_ctl_counts();
-}
-
-#[when("the Prose Pod API starts")]
-async fn when_api_starts(world: &mut TestWorld) {
-    assert!(world.api.is_none());
-    match test_server(world).await {
-        Ok(server) => world.api = Some(server),
-        Err(err) => {
-            world.api = None;
-            error!("Startup error: {err}");
-        }
-    }
-}
-
-#[then(expr = "the Prose Pod API {state_verb} running")]
-async fn then_api_run_state(world: &mut TestWorld, state: parameters::StateVerb) {
-    if state.into_bool() {
-        assert!(world.api.is_some());
-    } else {
-        assert!(world.api.is_none());
-    }
-}
-
-#[given("the XMPP server is offline")]
-fn given_xmpp_server_offline(world: &mut TestWorld) {
-    world.xmpp_service_state_mut().online = false;
-    world.server_ctl_state_mut().online = false;
-}
-
-#[given(expr = "the SMTP server {state_verb} reachable")]
-fn given_smtp_server_offline(world: &mut TestWorld, state: parameters::StateVerb) {
-    world.email_notifier_state_mut().online = state.into_bool();
-}
-
-#[then("the call should succeed")]
-fn then_response_success(world: &mut TestWorld) {
-    world.result().assert_status_success();
-}
-
-#[then("the call should not succeed")]
-#[then("the call should fail")]
-fn then_response_failure(world: &mut TestWorld) {
-    world.result().assert_status_failure();
-}
-
-#[then("the response body should be empty")]
-fn then_response_body_empty(world: &mut TestWorld) {
-    world.result().as_bytes().is_empty();
-}
-
-#[then("the response content type should be JSON")]
-fn then_content_type_json(world: &mut TestWorld) {
-    let res = world.result();
-    assert_eq!(
-        res.maybe_content_type(),
-        Some("application/json".to_owned())
-    );
-}
-
-#[then(expr = "the response JSON should contain key {string}")]
-fn then_response_json_has_key(world: &mut TestWorld, key: String) {
-    let map: serde_json::Map<String, serde_json::Value> = world.result().json();
-    assert!(map.keys().any(|k| k == &key), "{map:#?}")
-}
-
-#[then(expr = "the response JSON should not contain key {string}")]
-fn then_response_json_no_key(world: &mut TestWorld, key: String) {
-    let map: serde_json::Map<String, serde_json::Value> = world.result().json();
-    assert!(!map.keys().any(|k| k == &key), "{map:#?}")
-}
-
-#[then(expr = "the response content type should be {string}")]
-fn then_content_type(world: &mut TestWorld, content_type: String) {
-    let res = world.result();
-    assert_eq!(res.maybe_content_type(), Some(content_type));
-}
-
-#[then(expr = "the HTTP status code should be {status}")]
-fn then_response_http_status(world: &mut TestWorld, status: parameters::HTTPStatus) {
-    world.result().assert_status(*status);
-}
-
-#[then(expr = "the response should contain a {string} HTTP header")]
-fn then_response_headers_contain(world: &mut TestWorld, header_name: String) {
-    world.result().assert_contains_header(header_name);
-}
-
-#[then(expr = "the {string} header should contain {string}")]
-fn then_response_header_equals(world: &mut TestWorld, header_name: String, header_value: String) {
-    world.result().assert_header(header_name, header_value);
-}
-
-#[then("the response is a SSE stream")]
-fn then_response_is_sse_stream(world: &mut TestWorld) {
-    let res = world.result();
-    assert_eq!(
-        res.maybe_content_type(),
-        Some("text/event-stream".to_owned()),
-        "Content type (body: {:#?})",
-        res.text()
-    );
-}
-
-lazy_static! {
-    static ref SEMICOLON_SPACE_REGEX: Regex = Regex::new(r"(?m)(^.*:)\s").unwrap();
-}
-
-fn received_events(world: &mut TestWorld) -> Vec<HashSet<String>> {
-    (world.result().text())
-        .split("\n\n")
-        // Remove spaces after semicolons (`:`)
-        .map(|s| SEMICOLON_SPACE_REGEX.replace_all(&s, "$1").to_string())
-        .map(|s| s.lines().map(ToOwned::to_owned).collect::<HashSet<_>>())
-        .collect::<Vec<_>>()
-}
-
-#[then(expr = "one SSE with id {string} is")]
-async fn then_sse_event(world: &mut TestWorld, id: String, step: &Step) {
-    let value = step.docstring().unwrap().trim().to_owned();
-
-    let events = received_events(world);
-
-    let expected = value
-        // Unescape double quotes
-        .replace(r#"\""#, r#"""#)
-        // Unescape newlines
-        .replace("\\n", "\n");
-    let mut expected = (expected.lines())
-        .map(ToOwned::to_owned)
-        .collect::<HashSet<String>>();
-    expected.insert(format!("id:{id}"));
-
-    assert!(
-        events.iter().any(|set| set == &expected),
-        "events: {events:#?}\nexpected: {expected:#?}"
-    );
-}
-
-#[then(expr = "at least one SSE has id {string}")]
-async fn then_sse_event_id(world: &mut TestWorld, id: String) {
-    let events = received_events(world);
-    let expected = format!("id:{id}");
-
-    assert!(
-        events.iter().any(|set| set.contains(&expected)),
-        "events: {events:#?}\nexpected: {expected:#?}"
-    );
-}
-
-#[then(expr = "no SSE has id {string}")]
-async fn then_no_sse_event_id(world: &mut TestWorld, id: String) {
-    let events = received_events(world);
-    let expected = format!("id:{id}");
-
-    assert!(
-        !events.iter().any(|set| set.contains(&expected)),
-        "events: {events:#?}\nexpected: {expected:#?}"
-    );
-}
-
-#[then(expr = "<{jid}>'s password is changed")]
-fn then_password_changed(world: &mut TestWorld, jid: parameters::JID) {
-    assert_ne!(world.mock_secrets_store().changes_count(&jid), 0);
-}
