+++ src/rest-api/tests/prelude/test_api.rs
@@ -8,7 +8,7 @@ use prose_pod_api::{util::LifecycleManager, AppState, MinimalAppState};
 use service::xmpp::ServerCtlImpl as _;
 use tracing::*;
 
-use super::{steps::app_config::reload_config, test_world::TestWorld};
+use super::{app_config::reload_config, test_world::TestWorld};
 
 pub async fn test_server(world: &mut TestWorld) -> anyhow::Result<TestServer> {
     info!("Creating test router…");
